pyproject.toml
[project]
name = "stochasticstandsim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "plotly>=6.3.1",
    "stable-baselines3>=2.7.0",
    "streamlit>=1.50.0",
    "torch>=2.8.0",
    "pytest>=8.3.3",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "numpy>=2.3.3",
    "matplotlib>=3.10.7",
    "gymnasium>=1.2.1",
    "seaborn>=0.13.2",
    "dataclasses>=0.8",
]
